+++ NEWS.rst
@@ -17,6 +17,11 @@ Added
 Changed
 -------
 
+- Check whether Mopidy server is reachable: Relying only on network
+  availability lead to UI being unresponsive when Argos runs on the
+  same host as the server but network becomes unavailable `#181
+  <https://github.com/orontee/argos/issues/181>`_
+
 Removed
 -------